+++ CsvWinFormsApp/CsvWinFormsApp.csproj.user
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Compile Update="Forms\MainForm.cs">
+    <Compile Update="Forms\LoadingMessageForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="Forms\LoadingMessageForm.cs">
+    <Compile Update="Forms\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
   </ItemGroup>
